pyproject.toml
[project]
name = "wallet-coupon-creator-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "pillow>=11.0.0",
    "cryptography>=43.0.0",
    "httpx[http2]>=0.27.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.22",
]
